Make a successful request to the forecast.io API. This simply proves that we can read from the API. In this case we are getting a summary for the upcoming day or so of weather in human-language (English) terms.
Note that this commit exposes an API token linked to our forecast.io account.
This should not be a big deal given this is not highly publicized and anyways
no credit card is linked to our account. Later on this will be improved such
that e.g. our program will read the API Access Token from some Home Directory
file.