Create a Sparkline module
This is a module for creating Sparklines made of Text. The core function accepts a range and a list of values, which is visualized in data measures graphically supplied by eight particular unicode characters which show a vertical bar at varying heights.

Sparklines will be used under rows of numeric data to show their trend and relation to other dimensions.